Updated handling of nginx_site and nginx service
Removed the template for the tcr_datasets site from the nginx_site
resource as this resource (actually a definition) doesn't have the
notify capability needed to restart the service if the configuration is
updated.  The definition of `nginx_site` will restart, but won't if the
link into `sites-enabled` exists.  A big mess.

Also removed wrapper around the nginx service as I don't think it will
be necessary with this change.